Refactor for performance and modern C# features
Updated `Debug.ps1` to enhance `Connect-SonarrInstance` with `PassThru` and variable resolution. Refactored `RentedBuffer` to use `Span<T>` for improved memory efficiency. Simplified APIs in `RentedBuffer_T.cs` by renaming `Buffer` to `Span` and removing redundant methods.

Rewrote slicing in `CharSpanExtensions.cs` to use the range operator. Refactored `ObjectConverter.cs` for better readability and performance, including updates to `ConvertToObject`, `ReadNumber`, and `ReadObject`. Simplified boolean handling and string processing with `RentedBuffer`.

Optimized `AuthHandler.cs` by using `Span<byte>` for Base64 encoding. Updated `CharCollections.cs` and `Wildcard` methods to use `Span` and `RentedBuffer` for memory efficiency. Introduced a new `Enumerator` struct in `Wildcard` for efficient iteration. Changed `Wildcard.State` to support larger lengths and simplified the indexer API.

These changes improve code readability, maintainability, and performance by leveraging modern C# features and best practices.

+++ src/MG.Sonarr.Next/Strings/Wildcard-State.cs
@@ -6,14 +6,14 @@ public readonly partial struct Wildcard
 	private readonly struct State
 	{
 		private readonly WildcardMatchType _type;
-		private readonly ushort _length;
+		private readonly int _length;
 
 		public int Length => _length;
 		public WildcardMatchType Type => _type;
 
 		internal State(int length, WildcardMatchType type)
 		{
-			_length = (ushort)length;
+			_length = length;
 			_type = type;
 		}
